Minor suggested edits
These are suggested edits only. Mostly, they correct for a few typos, inconsistencies, and include
minor edits for clarity. (Aspects, for example, are now *Capitalized and Italicized* throughout.)
Further, this diff separate each sentence into its own, unwrapped line in `.rst` files. This may
look unruly in a text editor, but it has no effect on rendering, and it allows more readable diffs
in the future.